Add about 300 intel-tdx-module tasks and adjust existing ones to new structure
Instead of havocking memory with a for-loop and setting char values,
use the new __VERIFIER_nondet_memory.
havoc_object tasks are still using the fine-grained __VERIFIER_nondet_X methods
for primitive types.

These tasks use improved verification harnesses and the latest version of HarnessForge (1.2.0)

+++ c/intel-tdx-module/original/task-generation/formal/tdx_override/src/common/debug/tdx_debug.h
@@ -209,7 +209,7 @@ uint64_t td_debug_config(uint64_t leaf, uint64_t payload, uint64_t second_payloa
 
 //Debug Assertion
 #ifdef TDXFV_ENABLE_DEBUG_ASSERT
-#define tdx_debug_assert(cond) TDXFV_ASSERT(cond);
+#define tdx_debug_assert(cond) TDXFV_ASSERT((int) (cond));
 #else
 #define tdx_debug_assert(cond)
 #endif
